take, import: match stem in dict
This allows to take names like foo-01, foo-02 while only foo is in dict.

A name ending in a sequence of digits, optionally preceeded by a hyphen,
is split into its stem and suffix with the suffix being the hyphen/digit
part. The stem is basically a name not ending digits
which is then looked up in the names dict.

The following names would be looked up as both their respective name
and, failing that, as 'foo' in the names dict:

  - foo1
  - foo01
  - foo-01
  - foo-0001

However, names like 'foo-1-1', 'foo1-2' do not trigger stem lookup.